requirements.txt
Django==1.11
djangorestframework==3.6.2
pytz==2019.1
appdirs==1.4.3
packaging==16.8
pyparsing==2.2.0
six==1.10.0
uWSGI==2.0.15
